requirements.txt
SpeechRecognition==3.14.0
pyttsx3==2.92
pyaudio
wikipedia
google-generativeai
